partial code cleanup
