+++ frontend/src/app/(host)/listings/accommodations/[id]/edit/page.tsx
@@ -1,7 +1,7 @@
 "use server";
 import React from "react";
 import EditListingComponent from "./EditListingComponent";
-import { getAccommodation } from "@/app/utils/helpers/accommodationHelper";
+import { getAccommodation } from "@/app/utils/helpers/accommodation/request";
 
 interface EditListingPageProps {
   params: {
@@ -13,7 +13,7 @@ const EditListingPage: React.FC<EditListingPageProps> = async ({ params }) => {
   const listing = await getAccommodation(Number(params.id));
 
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between">
+    <main className="min-w-1/2 flex min-h-screen flex-col items-center justify-between">
       <EditListingComponent listing={listing} />
     </main>
   );
