+++ frontend/src/app/components/accommodation/Counter.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import PlusIcon from "../svgs/PlusIcon";
 import MinusIcon from "../svgs/MinusIcon";
 import type { Accommodation } from "@/app/interfaces/AccomodationData";
+import { Button } from "@nextui-org/react";
 
 interface CounterProps {
   name: string;
@@ -28,19 +29,27 @@ const Counter: React.FC<CounterProps> = ({ name, data, setData, id }) => {
   };
 
   return (
-    <div className="mt-5 flex w-full justify-between px-0">
-      <div>{name}</div>
+    <div className="my-1 flex w-full justify-between px-0">
+      <div className="flex items-center">{name}</div>
 
-      <div className="flex items-center p-2">
-        <div onClick={handleDecrement}>
+      <div className="grid grid-cols-3 items-center p-2">
+        <Button
+          className="flex justify-center rounded-full bg-white"
+          isIconOnly
+          onClick={handleDecrement}
+        >
           <MinusIcon />
-        </div>
+        </Button>
 
-        <div className="px-5">{data[id]}</div>
+        <div className="flex justify-center">{data[id]}</div>
 
-        <div onClick={handleIncrement}>
+        <Button
+          className="flex justify-center rounded-full bg-white"
+          isIconOnly
+          onClick={handleIncrement}
+        >
           <PlusIcon />
-        </div>
+        </Button>
       </div>
     </div>
   );
